feat(subscriptions): add API endpoints, yearly pricing, invoices and email
- Add yearly pricing to subscription plans seeder (~2 months discount)
- Create REST API endpoints for mobile/web agencies:
  GET  /subscriptions/plans     (public, list active plans)
  GET  /subscriptions/current   (auth, current subscription + stats)
  POST /subscriptions/subscribe (auth, initiate FedaPay payment)
  POST /subscriptions/cancel    (auth, cancel active subscription)
  GET  /subscriptions/history   (auth, paginated subscription history)
- Create SubscriptionPlanResource and SubscriptionResource
- Create SubscribeRequest form request with validation
- Create Invoice model and migration for subscription billing
- Create SubscriptionInvoiceMail with branded HTML email template
- Update SubscriptionService to generate invoice on activation
  and send invoice email to all agency members
- Fix Subscription::daysRemaining() float return type bug
- Add 12 Pest feature tests covering all subscription API endpoints

+++ app/Models/Subscription.php
@@ -139,7 +139,7 @@ class Subscription extends Model
             return 0;
         }
 
-        return max(0, now()->diffInDays($this->ends_at, false));
+        return (int) max(0, now()->diffInDays($this->ends_at, false));
     }
 
     /**
